ユーザー設定をAbstractionsに移動
